NIFI-8025 - OIDC moved to a separate class and tested to be working.
NIFI-8025 - Moved SAML to separate class and tested to be working.

NIFI-8025 - Added license to new files.

NIFI-8025 - Fixed a SAML logout consumer path.
